+++ test/shared/kaocha/test_util.clj
@@ -3,7 +3,8 @@
             [clojure.spec.alpha :as s]
             [expound.alpha :as expound]
             [kaocha.report :as report]
-            [kaocha.testable :as testable]))
+            [kaocha.testable :as testable]
+            [kaocha.output :as output]))
 
 (def ^:dynamic *report-history* nil)
 
@@ -35,3 +36,16 @@
          {:out (str o#)
           :err (str e#)
           :result r#}))))
+
+(defmacro expect-warning
+  {:style/indent [1]}
+  [pattern & body]
+  `(let [warnings# (atom [])]
+     (with-redefs [output/warn (fn [& xs#] (swap! warnings# conj xs#))]
+       (let [result# (do ~@body)]
+         (when-not (seq (filter #(re-find ~pattern (apply str %)) @warnings#))
+           (#'t/do-report {:type :fail
+                           :message (str "Expected test to generate a warning ("
+                                         ~pattern
+                                         ") but no warning occured.")}))
+         result#))))