+++ test/unit/kaocha/repl_test.clj
@@ -4,12 +4,16 @@
             [matcher-combinators.test]
             [kaocha.config :as config]))
 
-(is (match?
-     '{:kaocha/tests [{:kaocha.testable/id :foo
-                       :kaocha/test-paths ["test/foo"]}]
-       :kaocha/reporter [kaocha.report.progress/report]
-       :kaocha/color? false
-       :kaocha/fail-fast? true
-       :kaocha/plugins [:kaocha.plugin.alpha/xfail]}
+(deftest config-test
+  (is (match?
+       '{:kaocha/tests [{:kaocha.testable/id :foo
+                         :kaocha/test-paths ["test/foo"]}]
+         :kaocha/reporter [kaocha.report.progress/report]
+         :kaocha/color? false
+         :kaocha/fail-fast? true
+         :kaocha/plugins [:kaocha.plugin/randomize
+                          :kaocha.plugin/filter
+                          :kaocha.plugin/capture-output
+                          :kaocha.plugin.alpha/xfail]}
 
-     (repl/config {:config-file "fixtures/custom_config.edn"})))
+       (repl/config {:config-file "fixtures/custom_config.edn"}))))